CircleAPI/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.6.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distlib==0.3.6
Django==4.2.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.12.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
pandas==2.2.3
platformdirs==3.5.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.5.2
pydantic_core==2.20.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
requests==2.30.0
rich==13.7.1
shellingham==1.5.4
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.4.4
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.30.4
virtualenv==20.23.0
watchfiles==0.22.0
websockets==12.0
